pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "pillow>=11.2.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "qrcode>=8.2",
]
